@@ -8,11 +8,11 @@ readme = "README.md"
 homepage = "https://alacritty.org"
 repository = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.70.0"
+rust-version = "1.74.0"
 
 [dependencies.alacritty_terminal]
 path = "../alacritty_terminal"
-version = "0.23.0-dev"
+version = "0.24.1-dev"
 
 [dependencies.alacritty_config_derive]
 path = "../alacritty_config_derive"
@@ -27,20 +27,19 @@ ahash = { version = "0.8.6", features = ["no-rng"] }
 bitflags = "2.2.1"
 clap = { version = "4.2.7", features = ["derive", "env"] }
 copypasta = { version = "0.10.1", default-features = false }
-crossfont = { version = "0.7.0", features = ["force_system_fontconfig"] }
-glutin = { version = "0.31.1", default-features = false, features = ["egl", "wgl"] }
+crossfont = "0.8.0"
+glutin = { version = "0.32.0", default-features = false, features = ["egl", "wgl"] }
 home = "0.5.5"
 libc = "0.2"
 log = { version = "0.4", features = ["std", "serde"] }
 notify = "6.1.1"
 parking_lot = "0.12.0"
-raw-window-handle = "0.5"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 serde_yaml = "0.9.25"
 toml = "0.8.2"
 unicode-width = "0.1"
-winit = { version = "0.29.15", default-features = false, features = ["rwh_05", "serde"] }
+winit = { version = "0.30.4", default-features = false, features = ["rwh_06", "serde"] }
 
 [build-dependencies]
 gl_generator = "0.14.0"
@@ -60,7 +59,7 @@ objc = "0.2.2"
 
 [target.'cfg(windows)'.dependencies]
 dirs = "5.0.1"
-windows-sys = { version = "0.48", features = [
+windows-sys = { version = "0.52", features = [
     "Win32_UI_WindowsAndMessaging",
     "Win32_System_Threading",
     "Win32_System_Console",
