@@ -7,7 +7,7 @@ description = "Alacritty configuration abstractions"
 homepage = "https://alacritty.org"
 repository = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.70.0"
+rust-version = "1.74.0"
 
 [dependencies]
 log = { version = "0.4.17", features = ["serde"] }
