@@ -1,6 +1,6 @@
 [package]
 name = "alacritty_terminal"
-version = "0.23.0-dev"
+version = "0.24.1-dev"
 authors = ["Christian Duerr <contact@christianduerr.com>", "Joe Wilm <joe@jwilm.com>"]
 license = "Apache-2.0"
 description = "Library for writing terminal emulators"
@@ -8,7 +8,7 @@ readme = "../README.md"
 homepage = "https://alacritty.org"
 repository = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.70.0"
+rust-version = "1.74.0"
 
 [features]
 default = ["serde"]
@@ -34,7 +34,7 @@ signal-hook = "0.3.10"
 [target.'cfg(windows)'.dependencies]
 piper = "0.2.1"
 miow = "0.6.0"
-windows-sys = { version = "0.48.0", features = [
+windows-sys = { version = "0.52.0", features = [
     "Win32_System_Console",
     "Win32_Foundation",
     "Win32_Security",
