@@ -216,10 +216,10 @@ impl WindowIdentity {
     /// Override the [`WindowIdentity`]'s fields with the [`WindowOptions`].
     pub fn override_identity_config(&self, identity: &mut Identity) {
         if let Some(title) = &self.title {
-            identity.title = title.clone();
+            identity.title.clone_from(title);
         }
         if let Some(class) = &self.class {
-            identity.class = class.clone();
+            identity.class.clone_from(class);
         }
     }
 }
@@ -524,7 +524,7 @@ mod tests {
             let generated = String::from_utf8_lossy(&generated);
 
             let mut completion = String::new();
-            let mut file = File::open(format!("../extra/completions/{}", file)).unwrap();
+            let mut file = File::open(format!("../extra/completions/{file}")).unwrap();
             file.read_to_string(&mut completion).unwrap();
 
             assert_eq!(generated, completion);
