@@ -20,10 +20,10 @@ use winit::event::{
     ElementState, Modifiers, MouseButton, MouseScrollDelta, Touch as TouchEvent, TouchPhase,
 };
 #[cfg(target_os = "macos")]
-use winit::event_loop::EventLoopWindowTarget;
+use winit::event_loop::ActiveEventLoop;
 use winit::keyboard::ModifiersState;
 #[cfg(target_os = "macos")]
-use winit::platform::macos::EventLoopWindowTargetExtMacOS;
+use winit::platform::macos::ActiveEventLoopExtMacOS;
 use winit::window::CursorIcon;
 
 use alacritty_terminal::event::EventListener;
@@ -107,7 +107,7 @@ pub trait ActionContext<T: EventListener> {
     fn message(&self) -> Option<&Message>;
     fn config(&self) -> &UiConfig;
     #[cfg(target_os = "macos")]
-    fn event_loop(&self) -> &EventLoopWindowTarget<Event>;
+    fn event_loop(&self) -> &ActiveEventLoop;
     fn mouse_mode(&self) -> bool;
     fn clipboard_mut(&mut self) -> &mut Clipboard;
     fn scheduler_mut(&mut self) -> &mut Scheduler;
@@ -809,7 +809,7 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
         if self.ctx.terminal().mode().contains(TermMode::FOCUS_IN_OUT) {
             let chr = if is_focused { "I" } else { "O" };
 
-            let msg = format!("\x1b[{}", chr);
+            let msg = format!("\x1b[{chr}");
             self.ctx.write_to_pty(msg.into_bytes());
         }
     }
@@ -1004,17 +1004,18 @@ impl<T: EventListener, A: ActionContext<T>> Processor<T, A> {
         let mouse_bindings = self.ctx.config().mouse_bindings().to_owned();
 
         // If mouse mode is active, also look for bindings without shift.
-        let mut check_fallback = mouse_mode && mods.contains(ModifiersState::SHIFT);
+        let fallback_allowed = mouse_mode && mods.contains(ModifiersState::SHIFT);
+        let mut exact_match_found = false;
 
         for binding in &mouse_bindings {
             // Don't trigger normal bindings in mouse mode unless Shift is pressed.
-            if binding.is_triggered_by(mode, mods, &button) && (check_fallback || !mouse_mode) {
+            if binding.is_triggered_by(mode, mods, &button) && (fallback_allowed || !mouse_mode) {
                 binding.action.execute(&mut self.ctx);
-                check_fallback = false;
+                exact_match_found = true;
             }
         }
 
-        if check_fallback {
+        if fallback_allowed && !exact_match_found {
             let fallback_mods = mods & !ModifiersState::SHIFT;
             for binding in &mouse_bindings {
                 if binding.is_triggered_by(mode, fallback_mods, &button) {
@@ -1224,7 +1225,7 @@ mod tests {
         }
 
         #[cfg(target_os = "macos")]
-        fn event_loop(&self) -> &EventLoopWindowTarget<Event> {
+        fn event_loop(&self) -> &ActiveEventLoop {
             unimplemented!();
         }
 
@@ -1326,9 +1327,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Left,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::Click,
         input_delay: Duration::ZERO,
@@ -1342,9 +1343,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Right,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::Click,
         input_delay: Duration::ZERO,
@@ -1358,9 +1359,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Middle,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::Click,
         input_delay: Duration::ZERO,
@@ -1374,9 +1375,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Left,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::DoubleClick,
         input_delay: Duration::ZERO,
@@ -1390,9 +1391,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Left,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::Click,
         input_delay: CLICK_THRESHOLD,
@@ -1406,9 +1407,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Left,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id:  DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id:  WindowId::dummy(),
         },
         end_state: ClickState::TripleClick,
         input_delay: Duration::ZERO,
@@ -1422,9 +1423,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Left,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::Click,
         input_delay: CLICK_THRESHOLD,
@@ -1438,9 +1439,9 @@ mod tests {
             event: WindowEvent::MouseInput {
                 state: ElementState::Pressed,
                 button: MouseButton::Right,
-                device_id: unsafe { DeviceId::dummy() },
+                device_id: DeviceId::dummy(),
             },
-            window_id: unsafe { WindowId::dummy() },
+            window_id: WindowId::dummy(),
         },
         end_state: ClickState::Click,
         input_delay: Duration::ZERO,
