@@ -183,7 +183,7 @@ impl HintState {
     /// Update the alphabet used for hint labels.
     pub fn update_alphabet(&mut self, alphabet: &str) {
         if self.alphabet != alphabet {
-            self.alphabet = alphabet.to_owned();
+            alphabet.clone_into(&mut self.alphabet);
             self.keys.clear();
         }
     }
