@@ -111,7 +111,7 @@ fn find_socket(socket_path: Option<PathBuf>) -> IoResult<UnixStream> {
     if let Some(socket_path) = socket_path {
         // Ensure we inform the user about an invalid path.
         return UnixStream::connect(&socket_path).map_err(|err| {
-            let message = format!("invalid socket path {:?}", socket_path);
+            let message = format!("invalid socket path {socket_path:?}");
             IoError::new(err.kind(), message)
         });
     }
